+++ packages/screenshoter/CHANGELOG.md
@@ -4,6 +4,10 @@
 ## Unreleased
 
 
+## 1.0.2 - 2020/12/29
+
+- fix saveScreenshot
+- fix saveScreenshot
 ## 1.0.1 - 2020/12/1
 
 - round coordinate and size before image operations