+++ packages/visual-grid-client/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Unreleased
 
+
+## 15.4.1 - 2020/12/29
+
 - use `imagePositionInActiveFrame` instead of custom calculation
+- updated to @applitools/eyes-sdk-core@12.10.1 (from 12.10.0)
 
 ## 15.4.0 - 2020/12/18